chore: move jsoncons path code into dedicated files
It's only a code move, without functional changes.
This is in preparation to implementing the same path functionality
for flexbuffers objects.
