+++ src/core/json/detail/common.h
@@ -1,0 +1,13 @@
+// Copyright 2024, DragonflyDB authors.  All rights reserved.
+// See LICENSE for licensing terms.
+//
+
+#pragma once
+
+namespace dfly::json::detail {
+enum MatchStatus {
+  OUT_OF_BOUNDS,
+  MISMATCH,
+};
+
+}
